backend/requirements.txt
uvicorn==0.34.0
fastapi==0.115.8
SQLAlchemy
pydantic
passlib[bcrypt]  # If you have password hashing anywhere
python-dotenv
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
openai
 # Python's built-in email module
dnspython  # Sometimes required for certain email setups
apscheduler
requests  # If making HTTP requests (optional, but commonly needed)
boto3